removed modal dialog
