--- src/components/botiga/Soci.jsx
@@ -1,82 +1,0 @@
-import React, { useState, useEffect } from "react";
-import useAuthStore from "../../stores/useAuthStore";
-import useDataStore from "../../stores/useDataStore";
-import useCartStore from "../../stores/useCartStore";
-import { Navigate } from "react-router-dom";
-import ModalCard from "../../modals/ModalCard";
-import { useTranslation } from "react-i18next";
-import FullscreenSpinner from "../spinner/FullscreenSpinner";
-import Icon from "../ui/Icon";
-
-export default function SociDialog(props) {
-  const { onClose, open, dataRow = [], setProductData, loading } = props;
-  const { isLoggedIn } = useAuthStore();
-  const { membership = [] } = useDataStore();
-  const { addToCart } = useCartStore();
-  const [isSubscriber, setIsSubscriber] = useState(true);
-  const [redirect, setRedirect] = useState(false);
-  const [t] = useTranslation("translation");
-  const adaptedData = dataRow.length > 0 ? dataRow : membership[0];
-
-  const {
-    id,
-    price_range,
-    images,
-    description,
-    benefits = "",
-    has_stock,
-    discount = 0,
-    price,
-  } = adaptedData || {};
-
-  useEffect(() => {
-    // component unmount.
-    return () => {
-      setProductData && setProductData([]);
-    };
-  }, []);
-
-  const handleClose = () => {
-    onClose();
-  };
-
-  const handleAddClick = (id) => {
-    const selectedMembershipId =
-      membership.filter(function (soci) {
-        return soci.id === id;
-      })[0]?.variants[0] || {};
-    addToCart(selectedMembershipId);
-    onClose();
-    setRedirect(true);
-  };
-
-  if (redirect) {
-    return isLoggedIn ? null : <Navigate to="/login" replace />;
-  }
-
-  return loading ? (
-    <FullscreenSpinner />
-  ) : (
-    <ModalCard
-      handleClose={handleClose}
-      open={open}
-      handleAddClick={handleAddClick}
-      id={id}
-      title={`${t("banners.soci-curt")}!`}
-      price={price}
-      price_range={price_range}
-      imgArr={images}
-      buttonIcon={<Icon icon="shoppingCart" type="hoverable-black" />}
-      box1Title={t("modal.descripcio")}
-      box1Text={description}
-      box2Title={t("modal.beneficis")}
-      box2Text={benefits}
-      type="SOCI"
-      colorMode="dark"
-      isSubscriber={isSubscriber}
-      setIsSubscriber={setIsSubscriber}
-      has_stock={has_stock}
-      discount={discount}
-    />
-  );
-}
