+++ src/components/cardView/CardView.css
@@ -1,0 +1,230 @@
+.card-view {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-family: "Bebas Neue";
+}
+
+.card-view__header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-top: 20px;
+  max-width: 1060px;
+  width: 80%;
+}
+
+.card-view__title-box {
+  flex: 1;
+  min-width: 0;
+}
+
+.card-view__actions {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-left: 16px;
+  flex-shrink: 0;
+}
+
+.card-view__actions svg {
+  cursor: pointer;
+}
+
+.card-view__copied {
+  font-size: 14px;
+  font-family: "Montserrat", sans-serif;
+  white-space: nowrap;
+}
+
+.card-view__row {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  flex-wrap: wrap;
+  height: 100%;
+}
+
+.card-view__col-image {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  max-width: 580px;
+  min-width: 180px;
+  padding: 20px;
+  height: 100%;
+}
+
+.card-view__col-details {
+  display: flex;
+  flex-direction: column;
+  max-width: 480px;
+  min-width: 480px;
+  padding: 20px;
+  height: 100%;
+}
+
+@media (max-width: 542px) {
+  .card-view__col-image,
+  .card-view__col-details {
+    width: 100%;
+    max-width: inherit;
+    min-width: inherit;
+  }
+
+  .card-view__header {
+    width: 90%;
+  }
+}
+
+.card-view__description-title {
+  font-size: 1.6rem;
+  text-align: left;
+  text-transform: uppercase;
+  padding-top: 10px;
+  flex-grow: 1;
+}
+
+.card-view__description-content {
+  font-family: "Montserrat", sans-serif;
+  font-size: 1.1rem;
+  margin-top: -0.3rem;
+  margin-bottom: 0.5rem;
+  text-align: left;
+  flex-grow: 1;
+  line-height: initial;
+  white-space: pre-wrap;
+  overflow-wrap: break-word;
+  word-break: break-word;
+}
+
+/* Reused classes from ModalCardStyled.css for sub-components */
+
+.modal-card-location__row {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.6em;
+}
+
+.modal-card___title_small {
+  display: flex;
+  align-items: center;
+  font-family: "Bebas Neue";
+  font-size: 26px;
+  text-transform: uppercase;
+}
+
+.modal-card___title_small > svg {
+  margin-bottom: 5px;
+}
+
+.modal-card___title_small > span {
+  margin-left: 4px;
+}
+
+.modal-card___title_small a,
+.modal-card___title_small a:visited,
+.modal-card___title_small a:active {
+  color: var(--color-negro);
+  text-decoration: none;
+}
+
+.modal-card___title_small__tickets {
+  text-align: left;
+  line-height: 1.5rem;
+}
+
+.interactiveDataBox-activitat__row {
+  display: flex;
+  flex-grow: 1;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  align-items: center;
+  padding-right: 1rem;
+  width: 100%;
+  white-space: pre-wrap;
+}
+
+.interactiveDataBox-activitat__text-data {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  font-family: "Montserrat", sans-serif;
+  font-size: 20px;
+  color: var(--color-negro);
+}
+
+.interactiveDataBox-activitat__text-data > a {
+  color: var(--color-negro);
+}
+
+.interactiveDataBox-activitat__text-loca,
+.interactiveDataBox-activitat__text-loca-fullview {
+  text-align: left;
+}
+
+.interactiveDataBox-activitat__text-loca a,
+.interactiveDataBox-activitat__text-loca-fullview a {
+  color: inherit;
+  font-family: "Montserrat", sans-serif;
+  font-size: 20px;
+  line-height: 20px;
+}
+
+.interactiveDataBox-soci__row {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  width: 100%;
+  margin: 0.6rem 0rem;
+}
+
+.interactiveDataBox-soci__buttonBox {
+  display: flex;
+}
+
+.interactiveDataBox-product-sizes__button_active {
+  display: inline-flex;
+  background-color: var(--color-cream);
+  width: 2.188rem;
+  height: 2.188rem;
+  color: var(--color-negro);
+  border-width: 3px;
+  border-style: solid;
+  border-color: var(--color-negro);
+  font-size: 1.8rem;
+  margin-right: 0.15rem;
+  justify-content: center;
+  align-items: center;
+}
+
+.interactiveDataBox-product-sizes__button {
+  display: inline-flex;
+  background-color: var(--color-negro);
+  width: 2.188rem;
+  height: 2.188rem;
+  color: var(--color-cream);
+  font-size: 1.8rem;
+  margin-right: 0.15rem;
+  justify-content: center;
+  align-items: center;
+}
+
+.sizes {
+  cursor: pointer;
+}
+
+.modal-sizes-row {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+.modal-center-label {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 22px;
+}
