+++ src/components/forms/Register/RegisterLayout.jsx
@@ -1,22 +1,15 @@
 import React, { useState } from "react";
 import { Navigate, NavLink } from "react-router-dom";
-import SociDialog from "../../botiga/Soci";
 import RegisterForm from "./RegisterForm";
 import { useTranslation } from "react-i18next";
 import Icon from "../../ui/Icon";
-import "./RegisterLayout.css";
 import useProfileStore from "../../../stores/useProfileStore";
+import "./RegisterLayout.css";
 
 const RegisterLayout = (props) => {
   const [t] = useTranslation("translation");
   const { user_profile = "" } = useProfileStore();
   const [redirect, setRedirect] = useState(false);
-  const [open, setOpen] = React.useState(false);
-
-  const handleClick = () => {
-    setOpen(!open);
-  };
-
   const showLogin = () => {
     props.setViewState("login");
   };
@@ -29,7 +22,7 @@ const RegisterLayout = (props) => {
         <div className="logTitle">{t("login.registrat")}</div>
         {user_profile !== "LOGGED" && (
           <NavLink to="/memberships">
-            <div className="sociLogBanner" onClick={handleClick}>
+            <div className="sociLogBanner">
               <div>{t("login.encara")}</div>
               <div className="register-add-box">
                 <Icon icon="plus" type="orange" />
@@ -42,7 +35,6 @@ const RegisterLayout = (props) => {
         <div className="styled-link styled-link--bold" onClick={showLogin}>
           {`- ${t("login.inicia")} -`}
         </div>
-        {open && <SociDialog open={open} onClose={handleClick} />}
       </div>
     </div>
   );
