+++ src/components/layout/ActivitatsMainSection.jsx
@@ -1,53 +1,11 @@
-import React, { useState } from "react";
+import React from "react";
 import CardNew from "./CardNew";
-import axiosInstance from "../../axios";
-import { API_URL } from "../../utils/constants";
-import ActivitatDialog from "../../pages/agenda/components/Activitat";
+import { useNavigate } from "react-router-dom";
 import "./ActivitatsMainSection.css";
 
 const ActivitatsMainSection = (props) => {
   const { activitats } = props;
-  const [open, setOpen] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [eventData, setEventData] = useState([
-    {
-      id: 0,
-      name: "",
-      price: "",
-      images: [""],
-      discount: "",
-      datetime: "",
-      address: "",
-      description: "",
-      saved: "",
-      purchased: "",
-      stock: 0,
-      is_active: false,
-      artists: [false],
-    },
-  ]);
-
-  const fetchEvent = (data) => {
-    setLoading(true);
-    axiosInstance
-      .get(`${API_URL}events/${data.id}`, {})
-      .then((res) => {
-        setEventData(res?.data);
-        setLoading(false);
-      })
-      .then(handleClickOpen())
-      .catch((err) => {
-        setLoading(false);
-        console.warn("ERROR: ", err);
-      });
-  };
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
+  const navigate = useNavigate();
 
   return (
     <div className="activitats-main">
@@ -56,7 +14,7 @@ const ActivitatsMainSection = (props) => {
           <div
             key={data.name}
             className="activitats-main__card-item Card3GridItem"
-            onClick={() => fetchEvent(data)}
+            onClick={() => navigate(`/activitats/${data.id}`)}
           >
             <CardNew
               className="CardIndividual"
@@ -67,15 +25,6 @@ const ActivitatsMainSection = (props) => {
             />
           </div>
         ))}
-      {open && (
-        <ActivitatDialog
-          open={open}
-          dataRow={eventData}
-          setEventData={setEventData}
-          onClose={handleClose}
-          loading={loading}
-        />
-      )}
     </div>
   );
 };
