+++ src/pages/Botiga.jsx
@@ -1,64 +1,18 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import BotigaGeneral from "../components/botiga/BotigaGeneral";
 import useDataStore from "../stores/useDataStore";
 import ProductBanner from "../components/botiga/ProductBanner";
-import SociDialog from "../components/botiga/Soci";
 import PageLayout from "../components/layout/PageLayout/PageLayout";
-import { API_URL } from "../utils/constants";
 import { useTranslation } from "react-i18next";
-import { NavLink, useLocation } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 import { deleteStringDecimals } from "../utils/utils";
-import axiosInstance from "../axios";
 import useMediaQuery from "../hooks/use-media-query";
 
 function Botiga() {
-  const [open, setOpen] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [productData, setProductData] = useState([]);
   const { membership = [] } = useDataStore();
-  let location = useLocation();
-  const value = Object.fromEntries(new URLSearchParams(location.search));
-  const externalId = value?.id;
-  const id_soci = membership.map((x) => x.id);
-
-  const handleOpen = () => {
-    let arr = [];
-    setLoading(true);
-    axiosInstance
-      .get(`${API_URL}subscriptions/${id_soci[0]}`, {})
-      .then((resposta) => {
-        arr.push(resposta.data);
-        id_soci.length > 1 &&
-          axiosInstance
-            .get(`${API_URL}subscriptions/${id_soci[1]}`, {})
-            .then((res) => {
-              arr.push(res.data);
-            });
-      })
-      .then(() => {
-        setOpen(true);
-        setProductData(arr);
-        setLoading(false);
-      })
-      .catch((err) => {
-        setLoading(false);
-        console.warn("ERROR: ", err);
-      });
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
   const [t] = useTranslation("translation");
   const isMobile = useMediaQuery("(max-width:1163px)");
   const sociPreu = membership[0]?.price_range;
-
-  useEffect(() => {
-    if (externalId === "14" || externalId === "15") {
-      handleOpen();
-    }
-  }, []);
 
   return (
     <PageLayout
@@ -88,15 +42,6 @@ function Botiga() {
           />
         </NavLink>
       </div>
-      {open && (
-        <SociDialog
-          open={open}
-          onClose={handleClose}
-          dataRow={productData}
-          setProductData={setProductData}
-          loading={loading}
-        />
-      )}
       <div className="BotigaContent">
         <BotigaGeneral />
       </div>
