--- src/pages/external-events/ExternalEvents.css
@@ -1,14 +1,0 @@
-.external-events__box {
-  height: 100%;
-  background-color: var(--color-cream);
-}
-
-.external-events__error {
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  font-family: "Bebas Neue";
-  font-size: 30px;
-}
